Store pending action details to simplify HTML UI
And remove the start-versus-update distinction so that we're always
setting action details when we start an action, which we mostly need
to do anyway.

Since the HTML form needs to know the actual fields for each value.
I may also end up needing the ID of the room being added for the canvas
UI, but we'll see how that goes.